Fix links in nav startegies dropdown, bug in layout-strategy
The bug in layout-strategy.html was preventing all project items from
rendering on the corresponding strategy page. Confusingly, it would
only render a subset.

Brad’s fix was to move the variable assignment for the projects array
directly before the include statement that passes it as a parameter.

This works, but _why_ it works is still unclear to me, since I thought
as long as the array was declared inside the same scope (in this case,
the container div), it should work.